requirements.txt
beautifulsoup4==4.6.0
lxml==4.2.1
matplotlib==2.2.2
mysql-connector-python==8.0.11
